+++ Project/System/MyRTC.h
@@ -1,7 +1,7 @@
 #ifndef __MYRTC_H
 #define __MYRTC_H
 
-extern uint16_t MyRTC_Time[];
+extern int MyRTC_Time[];
 
 void MyRTC_Init(void);
 void MyRTC_SetTime(void);
